Add jwt, protected route

+++ backend/controllers/goalController.js
@@ -1,11 +1,12 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalSchema");
+const User = require("../models/userSchema");
 
 // @desc Gets All Goals
 // @route GET /api/goals
 // @access Private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find();
+  const goals = await Goal.find({ user: req.user._id });
   res.status(200).json({
     goals,
   });
@@ -21,6 +22,7 @@ const setGoal = asyncHandler(async (req, res) => {
   }
   const goal = await Goal.create({
     text: req.body.text,
+    user: req.user._id,
   });
   res.status(200).json({
     goal,
@@ -31,12 +33,21 @@ const setGoal = asyncHandler(async (req, res) => {
 // @route PUT /api/goals/:id
 // @access Private
 const updateGoal = asyncHandler(async (req, res) => {
-  console.log(req.params.id);
   const goal = await Goal.findById(req.params.id);
   if (!goal) {
     res.status(404);
     throw new Error("Goal not found");
   }
+  const user = await User.findById(req.user._id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (goal.user.toString() !== user._id.toString()) {
+    res.status(403); //
+    throw new Error("Not authorized");
+  }
+
   const updatedGoal = await Goal.findByIdAndUpdate(
     { _id: req.params.id },
     {
@@ -48,8 +59,6 @@ const updateGoal = asyncHandler(async (req, res) => {
       new: true,
     }
   );
-  console.log(goal);
-  console.log(updatedGoal);
   res.status(200).json({
     updatedGoal,
   });
@@ -64,6 +73,17 @@ const deleteGoal = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Goal not found");
   }
+
+  const user = await User.findById(req.user._id);
+  if (!user) {
+    res.status(401);
+    throw new Error("User not found");
+  }
+  if (goal.user.toString() !== user._id.toString()) {
+    res.status(403); //
+    throw new Error("Not authorized");
+  }
+
   await Goal.findByIdAndDelete(req.params.id);
   res.status(200).json({
     message: `Goal ${req.params.id} deleted successfully`,
